+++ CHANGELOG.md
@@ -55,6 +55,7 @@
 - Add shouldProcessScope
 - Add tests for shouldProcessScope and make sure it is called or the root
 - Add Scope.jsonDump
+- Add sourceNodesOnly option for jsonDump
 
 ### Changed